requirements.txt
streamlit==1.32.0
openai==1.17.0
langchain
langchain-openai==0.1.3
tavily-python==0.3.3
langchain-core
langchain-community
python-dotenv